+++ examples/BaseModel/SampleServer/Program.cs
@@ -1,6 +1,6 @@
-﻿#region Copyright (c) 2011-2022 Technosoftware GmbH. All rights reserved
+﻿#region Copyright (c) 2011-2024 Technosoftware GmbH. All rights reserved
 //-----------------------------------------------------------------------------
-// Copyright (c) 2021 Technosoftware GmbH. All rights reserved
+// Copyright (c) 2011-2024 Technosoftware GmbH. All rights reserved
 // Web: https://technosoftware.com 
 // 
 // License: 
@@ -25,7 +25,7 @@
 //
 // SPDX-License-Identifier: MIT
 //-----------------------------------------------------------------------------
-#endregion Copyright (c) 2011-2022 Technosoftware GmbH. All rights reserved
+#endregion Copyright (c) 2011-2024 Technosoftware GmbH. All rights reserved
 
 #region Using Directives
 using System;
@@ -36,8 +36,8 @@ using System.Threading.Tasks;
 using Mono.Options;
 
 using Opc.Ua;
+
 using Technosoftware.UaConfiguration;
-using Technosoftware.UaServer;
 using Technosoftware.UaServer.Sessions;
 using Technosoftware.UaBaseServer;
 #endregion
@@ -233,7 +233,7 @@ namespace SampleCompany.SampleServer
             ExitCode = ExitCode.Ok;
         }
         #endregion
-        
+
         private void OnCertificateValidation(CertificateValidator validator, CertificateValidationEventArgs e)
         {
             if (e.Error.StatusCode == StatusCodes.BadCertificateUntrusted)