+++ tutorials/SampleCompany/NodeManagers/Simulation/Namespaces.cs
@@ -1,0 +1,27 @@
+#region Copyright (c) 2022-2024 Technosoftware GmbH. All rights reserved
+//-----------------------------------------------------------------------------
+// Copyright (c) 2022-2024 Technosoftware GmbH. All rights reserved
+// Web: https://technosoftware.com 
+//
+// The Software is based on the OPC Foundation MIT License. 
+// The complete license agreement for that can be found here:
+// http://opcfoundation.org/License/MIT/1.00/
+//-----------------------------------------------------------------------------
+#endregion Copyright (c) 2022-2024 Technosoftware GmbH. All rights reserved
+
+#region Using Directives
+#endregion
+
+namespace SampleCompany.NodeManagers.Simulation
+{
+    /// <summary>
+    /// Defines constants for namespaces used by the servers.
+    /// </summary>
+    public static partial class Namespaces
+    {
+        /// <summary>
+        /// The namespace for the nodes provided by the Simulation server.
+        /// </summary>
+        public const string SimulationServer = "http://samplecompany.com/SampleServer/NodeManagers/Simulation";
+    }
+}